Split out tests.rs
